requirement.txt
beautifulsoup4==4.5.1
Django==1.11
mecab-python3==0.7
numpy==1.12.1
sklearn==0.18.1
